fix: address code review findings for issue #161
Extract SALT_ROUNDS to shared config/auth.js, move name arrays to
seeders/data/patron-names.js, align all dev passwords to welcome123,
replace blanket eslint-disable with line-specific comments, and use
QueryTypes.SELECT for idempotency check.
