+++ backend/src/config/auth.js
@@ -1,0 +1,3 @@
+const SALT_ROUNDS = 12;
+
+module.exports = { SALT_ROUNDS };
