+++ backend/src/seeders/data/patron-names.js
@@ -1,0 +1,209 @@
+// Curated first names (100 entries, diverse backgrounds)
+const FIRST_NAMES = [
+  'James',
+  'Maria',
+  'Robert',
+  'Patricia',
+  'Michael',
+  'Jennifer',
+  'William',
+  'Linda',
+  'David',
+  'Elizabeth',
+  'Richard',
+  'Barbara',
+  'Joseph',
+  'Susan',
+  'Thomas',
+  'Jessica',
+  'Christopher',
+  'Sarah',
+  'Charles',
+  'Karen',
+  'Daniel',
+  'Lisa',
+  'Matthew',
+  'Nancy',
+  'Anthony',
+  'Betty',
+  'Mark',
+  'Margaret',
+  'Donald',
+  'Sandra',
+  'Steven',
+  'Ashley',
+  'Andrew',
+  'Dorothy',
+  'Paul',
+  'Kimberly',
+  'Joshua',
+  'Emily',
+  'Kenneth',
+  'Donna',
+  'Kevin',
+  'Michelle',
+  'Brian',
+  'Carol',
+  'George',
+  'Amanda',
+  'Timothy',
+  'Melissa',
+  'Ronald',
+  'Deborah',
+  'Jason',
+  'Stephanie',
+  'Edward',
+  'Rebecca',
+  'Jeffrey',
+  'Sharon',
+  'Ryan',
+  'Laura',
+  'Jacob',
+  'Cynthia',
+  'Nicholas',
+  'Kathleen',
+  'Gary',
+  'Amy',
+  'Eric',
+  'Angela',
+  'Jonathan',
+  'Shirley',
+  'Stephen',
+  'Brenda',
+  'Larry',
+  'Emma',
+  'Justin',
+  'Anna',
+  'Scott',
+  'Pamela',
+  'Brandon',
+  'Nicole',
+  'Benjamin',
+  'Samantha',
+  'Samuel',
+  'Katherine',
+  'Raymond',
+  'Christine',
+  'Gregory',
+  'Helen',
+  'Frank',
+  'Debra',
+  'Alexander',
+  'Rachel',
+  'Patrick',
+  'Carolyn',
+  'Jack',
+  'Janet',
+  'Dennis',
+  'Catherine',
+  'Jerry',
+  'Heather',
+  'Tyler',
+  'Diane',
+];
+
+// Curated last names (100 entries, diverse backgrounds)
+const LAST_NAMES = [
+  'Smith',
+  'Johnson',
+  'Williams',
+  'Brown',
+  'Jones',
+  'Garcia',
+  'Miller',
+  'Davis',
+  'Rodriguez',
+  'Martinez',
+  'Hernandez',
+  'Lopez',
+  'Gonzalez',
+  'Wilson',
+  'Anderson',
+  'Thomas',
+  'Taylor',
+  'Moore',
+  'Jackson',
+  'Martin',
+  'Lee',
+  'Perez',
+  'Thompson',
+  'White',
+  'Harris',
+  'Sanchez',
+  'Clark',
+  'Ramirez',
+  'Lewis',
+  'Robinson',
+  'Walker',
+  'Young',
+  'Allen',
+  'King',
+  'Wright',
+  'Scott',
+  'Torres',
+  'Nguyen',
+  'Hill',
+  'Flores',
+  'Green',
+  'Adams',
+  'Nelson',
+  'Baker',
+  'Hall',
+  'Rivera',
+  'Campbell',
+  'Mitchell',
+  'Carter',
+  'Roberts',
+  'Gomez',
+  'Phillips',
+  'Evans',
+  'Turner',
+  'Diaz',
+  'Parker',
+  'Cruz',
+  'Edwards',
+  'Collins',
+  'Reyes',
+  'Stewart',
+  'Morris',
+  'Morales',
+  'Murphy',
+  'Cook',
+  'Rogers',
+  'Gutierrez',
+  'Ortiz',
+  'Morgan',
+  'Cooper',
+  'Peterson',
+  'Bailey',
+  'Reed',
+  'Kelly',
+  'Howard',
+  'Ramos',
+  'Kim',
+  'Cox',
+  'Ward',
+  'Richardson',
+  'Watson',
+  'Brooks',
+  'Chavez',
+  'Wood',
+  'James',
+  'Bennett',
+  'Gray',
+  'Mendoza',
+  'Ruiz',
+  'Hughes',
+  'Price',
+  'Alvarez',
+  'Castillo',
+  'Sanders',
+  'Patel',
+  'Myers',
+  'Long',
+  'Ross',
+  'Foster',
+  'Jimenez',
+];
+
+module.exports = { FIRST_NAMES, LAST_NAMES };
