+++ code-review-results/2026-04-03-issue-161.md
@@ -1,0 +1,41 @@
+# Code Review: Issue #161 — 2026-04-03
+
+## Findings
+
+| Severity | File | Issue | Pillar | Resolved |
+| -------- | ---- | ----- | ------ | -------- |
+| Medium | `backend/src/seeders/20260403120000-bulk-patrons.js:12` | Inconsistent dev password: bulk seeder uses `welcome123` while patron-passwords seeder uses `password123` — developers will face credential confusion | Documentation | No |
+| Medium | `backend/src/seeders/20260403120000-bulk-patrons.js:11` | `SALT_ROUNDS = 12` duplicated in 3 files (AuthService, patron-passwords seeder, bulk-patrons seeder) — should be a shared constant | Architecture | No |
+| Medium | `backend/src/seeders/20260403120000-bulk-patrons.js` | File is 323 lines, exceeding the 200-line limit. Name arrays (200+ lines of data) could be extracted to a separate data file | Architecture | No |
+| Low | `backend/src/seeders/20260403120000-bulk-patrons.js:274` | Raw SQL query `SELECT COUNT(*) as count FROM patrons` for idempotency check. Matches existing seeder conventions but violates the "no raw SQL" constraint in CLAUDE.md | Documentation | No |
+| Low | `backend/src/seeders/20260403120000-bulk-patrons.js:1` | `eslint-disable no-console` suppresses the structured-logger constraint. Acceptable for CLI-run seeders but follows an uncodified exception pattern | Documentation | No |
+
+## Harness Improvement Recommendations
+
+For each Medium+ finding, ask: **what harness change would prevent this class of issue from occurring in the first place?**
+
+| Finding | Harness Recommendation | Harness Change Made |
+| ------- | ---------------------- | ------------------- |
+| Inconsistent dev password | Add a CLAUDE.md constraint: "All dev/seed passwords must use the same value — define once in a shared config" | No |
+| Duplicated SALT_ROUNDS | Extract `SALT_ROUNDS` to a shared auth config constant (e.g., `backend/src/config/auth.js`) and add a CLAUDE.md constraint: "Auth constants must be imported from `config/auth.js`, never redefined locally" | No |
+| File length exceeds 200 lines | Add CLAUDE.md guidance: "Seeder data arrays should be extracted to separate data files when the seeder exceeds 200 lines, keeping logic and data separate" | No |
+
+## Harness Self-Audit
+
+Harness audit tool is not available in this project. Self-audit skipped.
+
+## Suggested E2E Tests
+
+Based on the changes in this review, the following end-to-end tests are recommended:
+
+| # | Scenario | Steps | Expected Result | Priority |
+|---|----------|-------|-----------------|----------|
+| 1 | Bulk patron seeder creates 5,000 records | Run `npm run db:seed`, then query patron count | Patron count increases by 5,000; all BULK-prefixed card numbers present from BULK-00001 to BULK-05000 | High |
+| 2 | Bulk patron seeder is idempotent | Run `npm run db:seed` twice consecutively | Second run skips seeding; patron count remains the same; no duplicate card numbers | High |
+| 3 | Bulk patron seeder rollback | Run seed, then run `npm run db:seed:undo`, verify removal | All BULK-prefixed patrons removed; demo patrons unaffected | Medium |
+| 4 | Bulk patron passwords are valid | Seed patrons, attempt login with BULK-00001 and dev password | Authentication succeeds with the seeded password | Medium |
+| 5 | Patron autocomplete handles large dataset | Seed 5,000 patrons, open patron search/autocomplete in frontend | Autocomplete remains responsive and returns paginated/filtered results | Medium |
+
+### Notes
+- The project uses `npm run db:seed` for seeding. E2E tests for seeders would require a test database and should verify both the `up` and `down` migrations.
+- Test #5 validates the original motivation for Issue #161 — ensuring the UI handles large patron datasets gracefully.
